+++ client/src/components/Workspace/index.tsx
@@ -2,7 +2,7 @@ import Mom from 'components/Mom';
 import Sidebar from 'components/Sidebar';
 import { useEffect, useState } from 'react';
 import { getWorkspaceInfo } from 'src/apis/workspace';
-import { MomContext } from 'src/contexts/mom';
+import { SelectedMomContext } from 'src/contexts/selected-mom';
 import { TMom } from 'src/types/mom';
 import { WorkspaceInfo } from 'src/types/workspace';
 
@@ -12,7 +12,7 @@ interface WorkspaceProps {
 
 function Workspace({ workspaceId }: WorkspaceProps) {
   const [workspace, setWorkspace] = useState<WorkspaceInfo | null>(null);
-  const [mom, setMom] = useState<TMom | null>(null);
+  const [selectedMom, setSelectedMom] = useState<TMom | null>(null);
 
   useEffect(() => {
     loadWorkspaceInfo();
@@ -25,17 +25,17 @@ function Workspace({ workspaceId }: WorkspaceProps) {
 
       // 제일 처음 입장했을 때 defalult 회의록 보여주기
       if (workspaceInfo.moms.length) {
-        setMom(workspaceInfo.moms[0]);
+        setSelectedMom(workspaceInfo.moms[0]);
       }
     }
   };
 
   return (
     workspace && (
-      <MomContext.Provider value={{ mom, setMom }}>
+      <SelectedMomContext.Provider value={{ selectedMom, setSelectedMom }}>
         <Sidebar workspace={workspace} />
         <Mom />
-      </MomContext.Provider>
+      </SelectedMomContext.Provider>
     )
   );
 }
