+++ client/src/contexts/selected-mom.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+import { TMom } from 'src/types/mom';
+
+interface IMomContext {
+  selectedMom: TMom | null;
+  setSelectedMom: React.Dispatch<React.SetStateAction<TMom | null>>;
+}
+
+export const SelectedMomContext = createContext<IMomContext | null>(null);
