+++ client/src/hooks/useSelectedMom.ts
@@ -1,8 +1,8 @@
 import { useContext } from 'react';
-import { MomContext } from 'src/contexts/mom';
+import { SelectedMomContext } from 'src/contexts/selected-mom';
 
 export default function useMom() {
-  const context = useContext(MomContext);
+  const context = useContext(SelectedMomContext);
 
   if (!context) throw new Error('아니요. 없어요.');
